Update bsky likes view test for lex refactor, minor fix
