--- packages/bsky/tests/views/votes.test.ts
@@ -1,257 +1,0 @@
-import AtpAgent, { AppBskyFeedGetPostThread } from '@atproto/api'
-import { SeedClient } from '../seeds/client'
-import votesSeed from '../seeds/votes'
-import {
-  CloseFn,
-  constantDate,
-  forSnapshot,
-  paginateAll,
-  processAll,
-  runTestServer,
-  TestServerInfo,
-} from '../_util'
-
-describe('pds vote views', () => {
-  let server: TestServerInfo
-  let agent: AtpAgent
-  let pdsAgent: AtpAgent
-  let close: CloseFn
-  let sc: SeedClient
-
-  // account dids, for convenience
-  let alice: string
-  let bob: string
-
-  beforeAll(async () => {
-    server = await runTestServer({
-      dbPostgresSchema: 'views_votes',
-    })
-    close = server.close
-    agent = new AtpAgent({ service: server.url })
-    pdsAgent = new AtpAgent({ service: server.pdsUrl })
-    sc = new SeedClient(pdsAgent)
-    await votesSeed(sc)
-    await processAll(server)
-    alice = sc.dids.alice
-    bob = sc.dids.bob
-  })
-
-  afterAll(async () => {
-    await close()
-  })
-
-  const getCursors = (items: { createdAt?: string }[]) =>
-    items.map((item) => item.createdAt ?? constantDate)
-
-  const getSortedCursors = (items: { createdAt?: string }[]) =>
-    getCursors(items).sort((a, b) => tstamp(b) - tstamp(a))
-
-  const tstamp = (x: string) => new Date(x).getTime()
-
-  it('fetches post votes', async () => {
-    const alicePost = await agent.api.app.bsky.feed.getVotes(
-      { uri: sc.posts[alice][1].ref.uriStr },
-      { headers: sc.getHeaders(alice, true) },
-    )
-
-    expect(forSnapshot(alicePost.data)).toMatchSnapshot()
-    expect(getCursors(alicePost.data.votes)).toEqual(
-      getSortedCursors(alicePost.data.votes),
-    )
-  })
-
-  it('fetches reply votes', async () => {
-    const bobReply = await agent.api.app.bsky.feed.getVotes(
-      { uri: sc.replies[bob][0].ref.uriStr },
-      { headers: sc.getHeaders(alice, true) },
-    )
-
-    expect(forSnapshot(bobReply.data)).toMatchSnapshot()
-    expect(getCursors(bobReply.data.votes)).toEqual(
-      getSortedCursors(bobReply.data.votes),
-    )
-  })
-
-  it('paginates', async () => {
-    const results = (results) => results.flatMap((res) => res.votes)
-    const paginator = async (cursor?: string) => {
-      const res = await agent.api.app.bsky.feed.getVotes(
-        {
-          uri: sc.posts[alice][1].ref.uriStr,
-          before: cursor,
-          limit: 2,
-        },
-        { headers: sc.getHeaders(alice, true) },
-      )
-      return res.data
-    }
-
-    const paginatedAll = await paginateAll(paginator)
-    paginatedAll.forEach((res) =>
-      expect(res.votes.length).toBeLessThanOrEqual(2),
-    )
-
-    const full = await agent.api.app.bsky.feed.getVotes(
-      { uri: sc.posts[alice][1].ref.uriStr },
-      { headers: sc.getHeaders(alice, true) },
-    )
-
-    expect(full.data.votes.length).toEqual(4)
-    expect(results(paginatedAll)).toEqual(results([full.data]))
-  })
-
-  it('filters by direction', async () => {
-    const full = await agent.api.app.bsky.feed.getVotes(
-      {
-        uri: sc.posts[alice][1].ref.uriStr,
-      },
-      { headers: sc.getHeaders(alice, true) },
-    )
-
-    const upvotes = await agent.api.app.bsky.feed.getVotes(
-      {
-        uri: sc.posts[alice][1].ref.uriStr,
-        direction: 'up',
-      },
-      { headers: sc.getHeaders(alice, true) },
-    )
-
-    const downvotes = await agent.api.app.bsky.feed.getVotes(
-      {
-        uri: sc.posts[alice][1].ref.uriStr,
-        direction: 'down',
-      },
-      { headers: sc.getHeaders(alice, true) },
-    )
-
-    expect(upvotes.data.votes.length).toEqual(2)
-    upvotes.data.votes.forEach((vote) => {
-      expect(vote.direction).toEqual('up')
-    })
-
-    expect(downvotes.data.votes.length).toEqual(2)
-    downvotes.data.votes.forEach((vote) => {
-      expect(vote.direction).toEqual('down')
-    })
-
-    // Upvote filter and downvote filter comprise all votes
-    const upAndDownvotes = upvotes.data.votes.concat(downvotes.data.votes)
-    expect(full.data.votes.length).toEqual(4)
-    full.data.votes.forEach((vote) => {
-      upAndDownvotes.some(
-        (v) => v.direction === vote.direction && v.actor.did === vote.actor.did,
-      )
-    })
-  })
-
-  describe('setVote()', () => {
-    it('sets and clears votes', async () => {
-      const asAlice = {
-        encoding: 'application/json',
-        headers: sc.getHeaders(alice),
-      } as const
-
-      let post: AppBskyFeedGetPostThread.OutputSchema
-      const getPost = async () => {
-        const result = await agent.api.app.bsky.feed.getPostThread(
-          {
-            uri: sc.posts[bob][0].ref.uriStr,
-            depth: 0,
-          },
-          { headers: sc.getHeaders(alice, true) },
-        )
-        return result.data
-      }
-
-      post = await getPost()
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost)
-          .downvoteCount,
-      ).toEqual(0)
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost).viewer,
-      ).toEqual({})
-
-      // Upvote
-      const { data: upvoted } = await pdsAgent.api.app.bsky.feed.setVote(
-        {
-          direction: 'up',
-          subject: {
-            uri: sc.posts[bob][0].ref.uriStr,
-            cid: sc.posts[bob][0].ref.cidStr,
-          },
-        },
-        asAlice,
-      )
-      await processAll(server)
-      post = await getPost()
-      expect(upvoted.upvote).not.toBeUndefined()
-      expect(upvoted.downvote).toBeUndefined()
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost)
-          .upvoteCount,
-      ).toEqual(1)
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost)
-          .downvoteCount,
-      ).toEqual(0)
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost).viewer,
-      ).toEqual(upvoted)
-
-      // Downvote
-      const { data: downvoted } = await pdsAgent.api.app.bsky.feed.setVote(
-        {
-          direction: 'down',
-          subject: {
-            uri: sc.posts[bob][0].ref.uriStr,
-            cid: sc.posts[bob][0].ref.cidStr,
-          },
-        },
-        asAlice,
-      )
-      await processAll(server)
-      post = await getPost()
-      expect(downvoted.upvote).toBeUndefined()
-      expect(downvoted.downvote).not.toBeUndefined()
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost)
-          .upvoteCount,
-      ).toEqual(0)
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost)
-          .downvoteCount,
-      ).toEqual(1)
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost).viewer,
-      ).toEqual(downvoted)
-
-      // No vote
-      const { data: novoted } = await pdsAgent.api.app.bsky.feed.setVote(
-        {
-          direction: 'none',
-          subject: {
-            uri: sc.posts[bob][0].ref.uriStr,
-            cid: sc.posts[bob][0].ref.cidStr,
-          },
-        },
-        asAlice,
-      )
-      await processAll(server)
-      post = await getPost()
-      expect(novoted.upvote).toBeUndefined()
-      expect(novoted.downvote).toBeUndefined()
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost)
-          .upvoteCount,
-      ).toEqual(0)
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost)
-          .downvoteCount,
-      ).toEqual(0)
-      expect(
-        (post.thread.post as AppBskyFeedGetPostThread.ThreadViewPost).viewer,
-      ).toEqual(novoted)
-    })
-  })
-})
